Update post dynamic page and PM buttons
