+++ apps/foundation/public/locales/en/common.json
@@ -8,8 +8,8 @@
   "apply-form.apply": "Apply Now",
   "apply-form.available-hours.input": "Weekly Available hours",
   "apply-form.available-hours.required": "You must specify number of your available hours",
-  "apply-form.city":"City",
-  "apply-form.city.required":"City field required",
+  "apply-form.city": "City",
+  "apply-form.city.required": "City field required",
   "apply-form.comment": "Comment",
   "apply-form.comment-placeholder": "Write Comment",
   "apply-form.comments": "Comments",
@@ -22,8 +22,8 @@
   "apply-form.in-mailing-list": "In mailing list",
   "apply-form.jobs.required": "You must select at least one job",
   "apply-form.jobs.title": "Jobs",
-  "apply-form.age.input":"Age",
-  "apply-form.age.required":"Age is required",
+  "apply-form.age.input": "Age",
+  "apply-form.age.required": "Age is required",
   "apply-form.name.input": "Name-surname",
   "apply-form.name.required": "Name field is required",
   "apply-form.occupation": "Occupation",
@@ -201,6 +201,7 @@
   "post.remaining-content": "Remaining Content",
   "post.remove": "Remove",
   "post.search-label": "Search Users",
+  "post.see-other-posts": "See Other Posts",
   "post.share-tweet": "Share Tweet",
   "post.tabs.archive": "Archive",
   "post.tabs.share": "Share",
