+++ apps/foundation/public/locales/nl/common.json
@@ -8,8 +8,8 @@
   "apply-form.apply": "Aanmelden Nu",
   "apply-form.available-hours.input": "Wekelijks beschikbare uren",
   "apply-form.available-hours.required": "U moet het aantal van uw beschikbare uren opgeven",
-  "apply-form.city":"Stad",
-  "apply-form.city.required":"Plaatsveld is verplicht",
+  "apply-form.city": "Stad",
+  "apply-form.city.required": "Plaatsveld is verplicht",
   "apply-form.comment": "Opmerking",
   "apply-form.comment-placeholder": "Schrijf Opmerking",
   "apply-form.comments": "Opmerkingen",
@@ -22,8 +22,8 @@
   "apply-form.in-mailing-list": "In mailinglijst",
   "apply-form.jobs.required": "U moet ten minste één baan selecteren",
   "apply-form.jobs.title": "Banen",
-  "apply-form.age.input":"Leeftijd",
-  "apply-form.age.required":"Leeftijd is verplicht",
+  "apply-form.age.input": "Leeftijd",
+  "apply-form.age.required": "Leeftijd is verplicht",
   "apply-form.name.input": "Voornaam Achternaam",
   "apply-form.name.required": "Naamveld is verplicht",
   "apply-form.occupation": "Beroep",
@@ -201,6 +201,7 @@
   "post.remaining-content": "Resterende Inhoud",
   "post.remove": "Verwijderen",
   "post.search-label": "Gebruikers Zoeken",
+  "post.see-other-posts": "Zie andere posts",
   "post.share-tweet": "Deel Tweet",
   "post.tabs.archive": "Archief",
   "post.tabs.share": "Delen",
