+++ apps/foundation/public/locales/tr/common.json
@@ -4,13 +4,13 @@
   "announcement.date": "Tarih: ",
   "announcement.title": "ETKİNLİK DUYURUSU",
   "announcement.topic": "Konu: ",
-  
+
   "apply-form.agreement": "<a>Kullanıcı sözleşmesini</a> okudum ve kabul ediyorum.",
   "apply-form.apply": "Şimdi Başvur",
   "apply-form.available-hours.input": "Haftalık müsait saat sayısı",
   "apply-form.available-hours.required": "Müsait saatlerinizin sayısını belirtmelisiniz",
-  "apply-form.city":"Şehir",
-  "apply-form.city.required":"Şehir alanı zorunludur",
+  "apply-form.city": "Şehir",
+  "apply-form.city.required": "Şehir alanı zorunludur",
   "apply-form.comment": "Yorum",
   "apply-form.comment-placeholder": "Yorum Yazın",
   "apply-form.comments": "Yorumlar",
@@ -26,8 +26,8 @@
   "apply-form.name.input": "Ad soyad",
   "apply-form.name.required": "Ad alanı zorunludur",
   "apply-form.occupation": "Meslek",
-  "apply-form.age.input":"Yaş",
-  "apply-form.age.required":"Yaş alanı zorunludur",
+  "apply-form.age.input": "Yaş",
+  "apply-form.age.required": "Yaş alanı zorunludur",
   "apply-form.phone.input": "Telefon",
   "apply-form.phone.required": "Telefon alanı zorunludur",
   "apply-form.send": "Gönder",
@@ -202,6 +202,7 @@
   "post.remaining-content": "Kalan İçerik",
   "post.remove": "Kaldır",
   "post.search-label": "Kullanıcı Ara",
+  "post.see-other-posts": "Diğer Postları Gör",
   "post.share-tweet": "Tweet Paylaş",
   "post.tabs.archive": "Arşiv",
   "post.tabs.share": "Paylaş",
