+++ apps/kunsthalte/public/locales/en/common.json
@@ -197,6 +197,7 @@
   "post.remaining-content": "Remaining Content",
   "post.remove": "Remove",
   "post.search-label": "Search Users",
+  "post.see-other-posts": "See Other Posts",
   "post.share-tweet": "Share Tweet",
   "post.tabs.archive": "Archive",
   "post.tabs.share": "Share",
