+++ apps/lotus/public/locales/nl/common.json
@@ -197,6 +197,7 @@
   "post.remaining-content": "Resterende Inhoud",
   "post.remove": "Verwijderen",
   "post.search-label": "Gebruikers Zoeken",
+  "post.see-other-posts": "Zie andere posts",
   "post.share-tweet": "Deel Tweet",
   "post.tabs.archive": "Archief",
   "post.tabs.share": "Delen",
