+++ apps/samenvvv/pages/hashtags/[slug]/index.tsx
@@ -1,30 +1,51 @@
-import { FC } from 'react'
+import { FC, useEffect } from 'react'
 
-import { useBreakpointValue } from '@chakra-ui/react'
+import {
+  Box,
+  Button,
+  Modal,
+  ModalBody,
+  ModalContent,
+  ModalFooter,
+  ModalOverlay,
+  Stack,
+  useBreakpointValue,
+  useDisclosure,
+} from '@chakra-ui/react'
 import { TourProvider } from '@reactour/tour'
 import { QueryClient, dehydrate } from '@tanstack/react-query'
 import { disableBodyScroll, enableBodyScroll } from 'body-scroll-lock'
 import { getCookie } from 'cookies-next'
 import { GetServerSidePropsContext, InferGetServerSidePropsType } from 'next'
+import Head from 'next/head'
+import { useTranslation } from 'next-i18next'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import { serialize } from 'next-mdx-remote/serialize'
 import { NextSeoProps } from 'next-seo'
 
+import { ASSETS_URL, SITE_URL } from '@wsvvrijheid/config'
 import {
   getHashtagBySlug,
   getHashtagSentences,
+  getModelById,
   useHashtag,
 } from '@wsvvrijheid/services'
-import { HashtagReturnType, StrapiLocale, Trend } from '@wsvvrijheid/types'
+import {
+  HashtagReturnType,
+  Post,
+  StrapiLocale,
+  Trend,
+} from '@wsvvrijheid/types'
 import {
   Container,
   HashtagProvider,
+  PostImage,
   PostMaker,
   StepsContent,
   TimeLeft,
   usePostMakerSteps,
 } from '@wsvvrijheid/ui'
-import { getPageSeo } from '@wsvvrijheid/utils'
+import { getItemLink, getOgImageSrc, getPageSeo } from '@wsvvrijheid/utils'
 
 import { Layout } from '../../../components'
 import i18nConfig from '../../../next-i18next.config'
@@ -34,10 +55,16 @@ type HashtagProps = InferGetServerSidePropsType<typeof getServerSideProps>
 const HashtagPage: FC<HashtagProps> = ({
   hasStarted,
   seo,
+  post,
+  capsSrc,
   isAdminMode,
   isIosSafari,
 }) => {
   const hashtag = useHashtag()
+
+  const { isOpen, onClose, onOpen } = useDisclosure()
+
+  const { t } = useTranslation()
 
   const isMobile = useBreakpointValue({ base: true, lg: false })
   const postMakerSteps = usePostMakerSteps()
@@ -46,6 +73,12 @@ const HashtagPage: FC<HashtagProps> = ({
     target && disableBodyScroll(target)
   const enableBody = (target: Element | null) =>
     target && enableBodyScroll(target)
+
+  useEffect(() => {
+    if (post) {
+      onOpen()
+    }
+  }, [post])
 
   if (!hashtag) return null
 
@@ -66,6 +99,27 @@ const HashtagPage: FC<HashtagProps> = ({
           }),
         }}
       >
+        {capsSrc && (
+          <Head>
+            <meta property="twitter:image:src" content={capsSrc} />
+          </Head>
+        )}
+        {post && (
+          <Modal isCentered isOpen={isOpen} onClose={onClose}>
+            <ModalOverlay />
+            <ModalContent>
+              <ModalBody p={0}>
+                <Stack>
+                  <PostImage size="sm" post={post} />
+                  <Box p={8}>{post.description}</Box>
+                </Stack>
+              </ModalBody>
+              <ModalFooter>
+                <Button onClick={onClose}>{t('post.see-other-posts')}</Button>
+              </ModalFooter>
+            </ModalContent>
+          </Modal>
+        )}
         <Layout seo={seo}>
           <Container py={4} pos="relative">
             {hasStarted || isAdminMode ? (
@@ -87,8 +141,82 @@ export const getServerSideProps = async (
 ) => {
   const locale = context.locale as StrapiLocale
   const slug = context.params?.slug as string
-  const { req, res } = context
+  const { req, res, query } = context
   const adminMode = getCookie('admin-mode', { req, res })
+
+  const post: Post = query.id
+    ? await getModelById<Post>({
+        url: 'api/posts',
+        id: Number(query.id),
+      })
+    : null
+
+  let seo: NextSeoProps = {}
+  let capsSrc: string
+
+  if (post) {
+    const title = post?.description?.slice(0, 20) || ''
+    const description = post.description || ''
+    const image = post?.image
+    const caps = post?.caps?.url
+
+    let src = image?.url
+    const link = getItemLink(post, locale, 'post') as string
+
+    if (caps) {
+      capsSrc = `${ASSETS_URL}${caps}`
+    } else {
+      if (image?.formats?.small) {
+        src = image.formats.small.url
+      } else if (image?.formats?.medium) {
+        src = image.formats.medium.url
+      } else if (image?.formats?.large) {
+        src = image.formats.large.url
+      }
+
+      capsSrc =
+        SITE_URL +
+        getOgImageSrc({
+          title: post.title,
+          text: post.description,
+          image: src ? `${ASSETS_URL}${src}` : undefined,
+          ...post.imageParams,
+        })
+    }
+
+    const images = image && [
+      {
+        url: capsSrc,
+        secureUrl: capsSrc,
+        type: image.mime as string,
+        width: 1200,
+        height: 675,
+        alt: title,
+      },
+    ]
+
+    const twitterHandle = {
+      en: '@samenvvvEn',
+      nl: '@samenvvv',
+      tr: '@samenvvvTr',
+    }
+
+    seo = {
+      title,
+      description,
+      twitter: {
+        cardType: 'summary_large_image',
+        site: twitterHandle[locale],
+        handle: twitterHandle[locale],
+      },
+      openGraph: {
+        title,
+        description,
+        url: link,
+        images,
+      },
+    }
+  }
 
   const userAgent = req.headers['user-agent']
   const isIOS = /iPad|iPhone|iPod/.test(userAgent)
@@ -126,7 +254,7 @@ export const getServerSideProps = async (
       { en: '', nl: '', tr: '' },
     ) || {}
 
-  const seo: NextSeoProps = getPageSeo(hashtag, locale, 'hashtag')
+  seo = getPageSeo(hashtag, locale, 'hashtag')
 
   const source = await serialize(hashtag?.content || '')
 
@@ -134,6 +262,8 @@ export const getServerSideProps = async (
     props: {
       source,
       seo,
+      capsSrc: capsSrc || null,
+      post: post || null,
       isIosSafari,
       isAdminMode: adminMode === true,
       slugs: { ...slugs, [locale]: slug },
