+++ apps/samenvvv/public/locales/tr/common.json
@@ -197,6 +197,7 @@
   "post.remaining-content": "Kalan İçerik",
   "post.remove": "Kaldır",
   "post.search-label": "Kullanıcı Ara",
+  "post.see-other-posts": "Diğer Postları Gör",
   "post.share-tweet": "Tweet Paylaş",
   "post.tabs.archive": "Arşiv",
   "post.tabs.share": "Paylaş",
