+++ packages/ui/public/locales/en/common.json
@@ -93,7 +93,7 @@
   "create-hashtag": "Create Hashtag",
   "create-news": "Create News",
   "create-post": "Create Post",
-  "create-multiple-post":"Create Multiple Post",
+  "create-multiple-post": "Create Multiple Post",
   "create-recommended-tweet": "Create Recommended Tweet",
   "date": "Date:",
   "date-end": "End Date",
@@ -198,6 +198,7 @@
   "post.remaining-content": "Remaining Content",
   "post.remove": "Remove",
   "post.search-label": "Search Users",
+  "post.see-other-posts": "See Other Posts",
   "post.share-tweet": "Share Tweet",
   "post.tabs.archive": "Archive",
   "post.tabs.share": "Share",
