+++ packages/ui/public/locales/nl/common.json
@@ -93,7 +93,7 @@
   "create-hashtag": "Creëren Hashtag",
   "create-news": "Creëren Nieuws",
   "create-post": "Creëren Post",
-  "create-multiple-post":"Creëren Multiple Post",
+  "create-multiple-post": "Creëren Multiple Post",
   "create-recommended-tweet": "Creëren Aanbevolen Tweet",
   "date": "Datum",
   "date-end": "Einddatum",
@@ -198,6 +198,7 @@
   "post.remaining-content": "Resterende Inhoud",
   "post.remove": "Verwijderen",
   "post.search-label": "Gebruikers Zoeken",
+  "post.see-other-posts": "Zie andere posts",
   "post.share-tweet": "Deel Tweet",
   "post.tabs.archive": "Archief",
   "post.tabs.share": "Delen",
