+++ packages/ui/public/locales/tr/common.json
@@ -93,7 +93,7 @@
   "create-hashtag": "Hashtag Oluştur",
   "create-news": "Haber Oluştur",
   "create-post": "Post Oluştur",
-  "create-multiple-post":"Çoklu Posta Oluştur",
+  "create-multiple-post": "Çoklu Posta Oluştur",
   "create-recommended-tweet": "Önerilen Tweet Oluştur",
   "date": "Tarih:",
   "date-end": "Bitiş",
@@ -198,6 +198,7 @@
   "post.remaining-content": "Kalan İçerik",
   "post.remove": "Kaldır",
   "post.search-label": "Kullanıcı Ara",
+  "post.see-other-posts": "Diğer Postları Gör",
   "post.share-tweet": "Tweet Paylaş",
   "post.tabs.archive": "Arşiv",
   "post.tabs.share": "Paylaş",
